refactor: split compiler mod.rs into submodules
