+++ src/compiler/mod.rs
@@ -1,7 +1,6 @@
 use crate::models::{
-    ArkadeCovenant, AssetLookupSource, CompilerInfo, ContractJson, Expression, Function,
-    FunctionInput, GroupIOSource, GroupSumSource, Parameter, Requirement, Statement,
-    DEFAULT_ARRAY_LENGTH,
+    ArkadeCovenant, CompilerInfo, ContractJson, Expression, Function, FunctionInput, Parameter,
+    Requirement, Statement, DEFAULT_ARRAY_LENGTH,
 };
 use crate::opcodes::{
     OP_0, OP_1, OP_ADD64, OP_BIN2NUM, OP_BOOLAND, OP_CAT, OP_CHECKLOCKTIMEVERIFY, OP_CHECKSIG,
@@ -24,11 +23,27 @@ use crate::opcodes::{
     OP_SUBSTR, OP_SWAP, OP_TWEAKVERIFY, OP_TXHASH, OP_TXID, OP_TXWEIGHT, OP_VERIFY,
 };
 use crate::parser;
-use crate::typechecker::{self, ArkType};
+use crate::typechecker::{self};
 use crate::validator::{self, Severity};
 use chrono::Utc;
 
 pub mod tapscript;
+
+// ASM codegen and rewrite passes split into submodules;
+// siblings reach each other via `use super::*`.
+mod asset;
+mod comparison;
+mod concat;
+mod expr;
+mod introspection;
+mod loops;
+
+pub(crate) use asset::*;
+pub(crate) use comparison::*;
+pub(crate) use concat::*;
+pub(crate) use expr::*;
+pub(crate) use introspection::*;
+pub(crate) use loops::*;
 
 fn strip_comments(source: &str) -> String {
     source
@@ -341,1897 +356,6 @@ fn generate_requirement_asm(req: &Requirement, asm: &mut Vec<String>) -> Result<
         Requirement::Comparison { left, op, right } => {
             generate_comparison_asm(left, op, right, asm);
             Ok(())
-        }
-    }
-}
-
-/// Generate assembly for expression (for use in if conditions)
-fn generate_expression_asm(expr: &Expression, asm: &mut Vec<String>) {
-    match expr {
-        Expression::Variable(var) => {
-            asm.push(format!("<{}>", var));
-        }
-        Expression::Literal(lit) => {
-            asm.push(lit.clone());
-        }
-        Expression::Property(prop) => {
-            // Map the introspector "this" properties to their dedicated opcodes
-            // (the parser stores them as Property strings; resolving them here
-            // keeps the placeholder pipeline untouched for everything else).
-            match prop.as_str() {
-                "this.activeInputIndex" => asm.push(OP_PUSHCURRENTINPUTINDEX.to_string()),
-                "this.activeBytecode" => asm.push(OP_INPUTBYTECODE.to_string()),
-                _ => asm.push(format!("<{}>", prop)),
-            }
-        }
-        Expression::BinaryOp { left, op, right } => {
-            // Emit left operand
-            generate_expression_asm(left, asm);
-
-            // Convert to u64le if needed (witness inputs arrive as CScriptNum)
-            if needs_u64_conversion(left) {
-                asm.push(OP_SCRIPTNUMTOLE64.to_string());
-            }
-
-            // Emit right operand
-            generate_expression_asm(right, asm);
-
-            // Convert to u64le if needed
-            if needs_u64_conversion(right) {
-                asm.push(OP_SCRIPTNUMTOLE64.to_string());
-            }
-
-            // Emit opcode with OP_VERIFY for 64-bit ops (same as emit_binary_op_asm)
-            match op.as_str() {
-                "+" => {
-                    asm.push(OP_ADD64.to_string());
-                    asm.push(OP_VERIFY.to_string());
-                }
-                "-" => {
-                    asm.push(OP_SUB64.to_string());
-                    asm.push(OP_VERIFY.to_string());
-                }
-                "*" => {
-                    asm.push(OP_MUL64.to_string());
-                    asm.push(OP_VERIFY.to_string());
-                }
-                "/" => {
-                    asm.push(OP_DIV64.to_string());
-                    asm.push(OP_VERIFY.to_string());
-                }
-                ">=" => {
-                    asm.push(OP_GREATERTHANOREQUAL64.to_string());
-                    asm.push(OP_VERIFY.to_string());
-                }
-                "<=" => {
-                    asm.push(OP_LESSTHANOREQUAL64.to_string());
-                    asm.push(OP_VERIFY.to_string());
-                }
-                ">" => {
-                    asm.push(OP_GREATERTHAN64.to_string());
-                    asm.push(OP_VERIFY.to_string());
-                }
-                "<" => {
-                    asm.push(OP_LESSTHAN64.to_string());
-                    asm.push(OP_VERIFY.to_string());
-                }
-                "==" => asm.push(OP_EQUAL.to_string()),
-                "!=" => {
-                    asm.push(OP_EQUAL.to_string());
-                    asm.push(OP_NOT.to_string());
-                }
-                _ => asm.push(OP_FALSE.to_string()),
-            }
-        }
-        Expression::CurrentInput(property) => {
-            if let Some(prop) = property {
-                match prop.as_str() {
-                    "scriptPubKey" => asm.push(OP_INPUTBYTECODE.to_string()),
-                    "value" => asm.push(OP_INPUTVALUE.to_string()),
-                    "sequence" => asm.push(OP_INPUTSEQUENCE.to_string()),
-                    "outpoint" => asm.push(OP_INPUTOUTPOINT.to_string()),
-                    _ => asm.push(OP_INPUTBYTECODE.to_string()),
-                }
-            } else {
-                asm.push(OP_INPUTBYTECODE.to_string());
-            }
-        }
-        Expression::CheckSigExpr { signature, pubkey } => {
-            asm.push(format!("<{}>", pubkey));
-            asm.push(format!("<{}>", signature));
-            asm.push(OP_CHECKSIG.to_string());
-        }
-        Expression::CheckSigFromStackExpr {
-            signature,
-            pubkey,
-            message,
-        } => {
-            asm.push(format!("<{}>", message));
-            asm.push(format!("<{}>", pubkey));
-            asm.push(format!("<{}>", signature));
-            asm.push(OP_CHECKSIGFROMSTACK.to_string());
-        }
-        // Streaming SHA256
-        Expression::Sha256 { data } => {
-            generate_expression_asm(data, asm);
-            asm.push(OP_SHA256.to_string());
-        }
-        Expression::Sha256Initialize { data } => {
-            generate_expression_asm(data, asm);
-            asm.push(OP_SHA256INITIALIZE.to_string());
-        }
-        Expression::Sha256Update { context, chunk } => {
-            generate_expression_asm(context, asm);
-            generate_expression_asm(chunk, asm);
-            asm.push(OP_SHA256UPDATE.to_string());
-        }
-        Expression::Sha256Finalize {
-            context,
-            last_chunk,
-        } => {
-            generate_expression_asm(context, asm);
-            generate_expression_asm(last_chunk, asm);
-            asm.push(OP_SHA256FINALIZE.to_string());
-        }
-        // Byte-string concatenation: bytes-like + value → OP_CAT
-        // Coercion flags decide whether to convert a scriptnum side to LE64
-        // before the cat (so off-chain hashing uses fixed 8-byte ints).
-        Expression::Concat {
-            left,
-            right,
-            coerce_left,
-            coerce_right,
-        } => {
-            generate_expression_asm(left, asm);
-            if *coerce_left {
-                asm.push(OP_SCRIPTNUMTOLE64.to_string());
-            }
-            generate_expression_asm(right, asm);
-            if *coerce_right {
-                asm.push(OP_SCRIPTNUMTOLE64.to_string());
-            }
-            asm.push(OP_CAT.to_string());
-        }
-        // Conversion & Arithmetic
-        Expression::Neg64 { value } => {
-            generate_expression_asm(value, asm);
-            asm.push(OP_NEG64.to_string());
-        }
-        Expression::Le64ToScriptNum { value } => {
-            generate_expression_asm(value, asm);
-            asm.push(OP_LE64TOSCRIPTNUM.to_string());
-        }
-        Expression::Le32ToLe64 { value } => {
-            generate_expression_asm(value, asm);
-            asm.push(OP_LE32TOLE64.to_string());
-        }
-        // Crypto Opcodes
-        Expression::EcMulScalarVerify {
-            scalar,
-            point_p,
-            point_q,
-        } => {
-            generate_expression_asm(point_q, asm);
-            generate_expression_asm(point_p, asm);
-            generate_expression_asm(scalar, asm);
-            asm.push(OP_ECMULSCALARVERIFY.to_string());
-        }
-        Expression::TweakVerify {
-            point_p,
-            tweak,
-            point_q,
-        } => {
-            generate_expression_asm(point_q, asm);
-            generate_expression_asm(tweak, asm);
-            generate_expression_asm(point_p, asm);
-            asm.push(OP_TWEAKVERIFY.to_string());
-        }
-        Expression::CheckSigFromStackVerify {
-            signature,
-            pubkey,
-            message,
-        } => {
-            asm.push(format!("<{}>", message));
-            asm.push(format!("<{}>", pubkey));
-            asm.push(format!("<{}>", signature));
-            asm.push(OP_CHECKSIGFROMSTACKVERIFY.to_string());
-        }
-        Expression::AssetLookup {
-            source,
-            index,
-            asset_txid,
-            asset_gidx,
-        } => {
-            emit_asset_lookup_asm(source, index, asset_txid, asset_gidx, asm);
-        }
-        Expression::AssetHas {
-            source,
-            index,
-            asset_txid,
-            asset_gidx,
-        } => {
-            emit_asset_has_asm(source, index, asset_txid, asset_gidx, asm);
-        }
-        Expression::AssetCount { source, index } => {
-            emit_asset_count_asm(source, index, asm);
-        }
-        Expression::AssetAt {
-            source,
-            io_index,
-            asset_index,
-            property,
-        } => {
-            emit_asset_at_asm(source, io_index, asset_index, property, asm);
-        }
-        Expression::TxIntrospection { property } => {
-            emit_tx_introspection_asm(property, asm);
-        }
-        Expression::InputIntrospection { index, property } => {
-            emit_input_introspection_asm(index, property, asm);
-        }
-        Expression::OutputIntrospection { index, property } => {
-            emit_output_introspection_asm(index, property, asm);
-        }
-        Expression::GroupFind {
-            asset_txid,
-            asset_gidx,
-        } => {
-            emit_group_find_asm(asset_txid, asset_gidx, asm);
-        }
-        Expression::GroupHas {
-            asset_txid,
-            asset_gidx,
-        } => {
-            emit_group_has_asm(asset_txid, asset_gidx, asm);
-        }
-        Expression::GroupControlIs {
-            group,
-            asset_txid,
-            asset_gidx,
-        } => {
-            emit_group_control_is_asm(group, asset_txid, asset_gidx, asm);
-        }
-        Expression::GroupProperty { group, property } => {
-            emit_group_property_asm(group, property, asm);
-        }
-        Expression::AssetGroupsLength => {
-            asm.push(OP_INSPECTNUMASSETGROUPS.to_string());
-        }
-        Expression::GroupSum { index, source } => {
-            generate_expression_asm(index, asm);
-            match source {
-                GroupSumSource::Inputs => asm.push(OP_0.to_string()),
-                GroupSumSource::Outputs => asm.push(OP_1.to_string()),
-            }
-            asm.push(OP_INSPECTASSETGROUPSUM.to_string());
-        }
-        Expression::GroupNumIO { index, source } => {
-            generate_expression_asm(index, asm);
-            match source {
-                GroupIOSource::Inputs => asm.push(OP_0.to_string()),
-                GroupIOSource::Outputs => asm.push(OP_1.to_string()),
-            }
-            asm.push(OP_INSPECTASSETGROUPNUM.to_string());
-        }
-        Expression::GroupIOAccess {
-            group_index,
-            io_index,
-            source,
-            property,
-        } => {
-            generate_expression_asm(group_index, asm);
-            generate_expression_asm(io_index, asm);
-            match source {
-                GroupIOSource::Inputs => asm.push(OP_0.to_string()),
-                GroupIOSource::Outputs => asm.push(OP_1.to_string()),
-            }
-            asm.push(OP_INSPECTASSETGROUP.to_string());
-            // Extract property if specified
-            // Stack after opcode: type_u8, data..., amount_u64 (top)
-            if let Some(prop) = property {
-                match prop.as_str() {
-                    "amount" => {
-                        // Keep only amount (top of stack)
-                        // Need to handle based on type, but for now just keep top
-                    }
-                    "type" => {
-                        // Drop everything except type
-                        asm.push(OP_DROP.to_string()); // amount
-                        asm.push(OP_DROP.to_string()); // data (varies)
-                    }
-                    _ => {}
-                }
-            }
-        }
-        Expression::ContractInstance {
-            contract_name,
-            args,
-        } => {
-            emit_contract_instance_asm(contract_name, args, asm);
-        }
-        // Byte-string manipulation (introspector extensions)
-        Expression::Substr { data, offset, size } => {
-            generate_expression_asm(data, asm);
-            generate_expression_asm(offset, asm);
-            generate_expression_asm(size, asm);
-            asm.push(OP_SUBSTR.to_string());
-        }
-        Expression::Cat { left, right } => {
-            generate_expression_asm(left, asm);
-            generate_expression_asm(right, asm);
-            asm.push(OP_CAT.to_string());
-        }
-        Expression::Bin2Num { data } => {
-            generate_expression_asm(data, asm);
-            asm.push(OP_BIN2NUM.to_string());
-        }
-        Expression::Num2Bin { value, size } => {
-            generate_expression_asm(value, asm);
-            generate_expression_asm(size, asm);
-            asm.push(OP_NUM2BIN.to_string());
-        }
-        Expression::SizeOf { data } => {
-            generate_expression_asm(data, asm);
-            // OP_SIZE pushes len next to original bytes; OP_NIP drops the
-            // original so only the size remains on the stack.
-            asm.push(OP_SIZE.to_string());
-            asm.push(OP_NIP.to_string());
-        }
-        // Packet introspection
-        Expression::PacketInspect { packet_type } => {
-            generate_expression_asm(packet_type, asm);
-            asm.push(OP_INSPECTPACKET.to_string());
-            // OP_INSPECTPACKET returns (content, 1) on hit, () then 0 on miss.
-            // Assert present and discard the bool, leaving content on the stack.
-            asm.push(OP_1.to_string());
-            asm.push(OP_EQUALVERIFY.to_string());
-        }
-        Expression::InputPacketInspect { index, packet_type } => {
-            generate_expression_asm(packet_type, asm);
-            generate_expression_asm(index, asm);
-            asm.push(OP_INSPECTINPUTPACKET.to_string());
-            asm.push(OP_1.to_string());
-            asm.push(OP_EQUALVERIFY.to_string());
-        }
-    }
-}
-
-/// Generate assembly for comparison expressions
-fn generate_comparison_asm(left: &Expression, op: &str, right: &Expression, asm: &mut Vec<String>) {
-    match (left, op, right) {
-        (Expression::Variable(var), ">=", Expression::Literal(value)) => {
-            asm.push(format!("<{}>", var));
-            asm.push(value.clone());
-            asm.push(OP_GREATERTHANOREQUAL.to_string());
-        }
-        (Expression::Variable(var), "==", Expression::Variable(var2)) => {
-            asm.push(format!("<{}>", var));
-            asm.push(format!("<{}>", var2));
-            asm.push(OP_EQUAL.to_string());
-        }
-        (Expression::Variable(var), ">=", Expression::Variable(var2)) => {
-            asm.push(format!("<{}>", var));
-            asm.push(format!("<{}>", var2));
-            asm.push(OP_GREATERTHANOREQUAL.to_string());
-        }
-        (Expression::Variable(var), "==", Expression::Property(prop)) => {
-            asm.push(format!("<{}>", var));
-            asm.push(format!("<{}>", prop));
-            asm.push(OP_EQUAL.to_string());
-        }
-        (Expression::Variable(var), ">=", Expression::Property(prop)) => {
-            asm.push(format!("<{}>", var));
-            asm.push(format!("<{}>", prop));
-            asm.push(OP_GREATERTHANOREQUAL.to_string());
-        }
-        (Expression::Literal(lit), "==", Expression::Variable(var)) => {
-            asm.push(lit.clone());
-            asm.push(format!("<{}>", var));
-            asm.push(OP_EQUAL.to_string());
-        }
-        (Expression::Literal(lit), ">=", Expression::Variable(var)) => {
-            asm.push(lit.clone());
-            asm.push(format!("<{}>", var));
-            asm.push(OP_GREATERTHANOREQUAL.to_string());
-        }
-        (Expression::Literal(lit), "==", Expression::Literal(value)) => {
-            asm.push(lit.clone());
-            asm.push(value.clone());
-            asm.push(OP_EQUAL.to_string());
-        }
-        (Expression::Literal(lit), ">=", Expression::Literal(value)) => {
-            asm.push(lit.clone());
-            asm.push(value.clone());
-            asm.push(OP_GREATERTHANOREQUAL.to_string());
-        }
-        (Expression::Literal(lit), "==", Expression::Property(prop)) => {
-            asm.push(lit.clone());
-            asm.push(format!("<{}>", prop));
-            asm.push(OP_EQUAL.to_string());
-        }
-        (Expression::Literal(lit), ">=", Expression::Property(prop)) => {
-            asm.push(lit.clone());
-            asm.push(format!("<{}>", prop));
-            asm.push(OP_GREATERTHANOREQUAL.to_string());
-        }
-        (Expression::Property(prop), "==", Expression::Variable(var)) => {
-            asm.push(format!("<{}>", prop));
-            asm.push(format!("<{}>", var));
-            asm.push(OP_EQUAL.to_string());
-        }
-        (Expression::Property(prop), ">=", Expression::Variable(var)) => {
-            asm.push(format!("<{}>", prop));
-            asm.push(format!("<{}>", var));
-            asm.push(OP_GREATERTHANOREQUAL.to_string());
-        }
-        (Expression::Property(prop), "==", Expression::Literal(value)) => {
-            // Map "this" properties to their dedicated opcodes (mirrors
-            // generate_expression_asm); otherwise keep the placeholder.
-            let emit_left = |asm: &mut Vec<String>| match prop.as_str() {
-                "this.activeInputIndex" => asm.push(OP_PUSHCURRENTINPUTINDEX.to_string()),
-                "this.activeBytecode" => asm.push(OP_INPUTBYTECODE.to_string()),
-                _ => asm.push(format!("<{}>", prop)),
-            };
-            if value == "true" {
-                // `require(expr)` is parsed as `expr == true`; this dummy
-                // comparison just pushes the introspection result.
-                emit_left(asm);
-            } else {
-                // Correct Bitcoin Script order is left, right, OP_EQUAL.
-                emit_left(asm);
-                asm.push(value.clone());
-                asm.push(OP_EQUAL.to_string());
-            }
-        }
-        (Expression::Property(prop), ">=", Expression::Literal(value)) => {
-            asm.push(format!("<{}>", prop));
-            asm.push(value.clone());
-            asm.push(OP_GREATERTHANOREQUAL.to_string());
-        }
-        (Expression::Property(prop), "==", Expression::Property(prop2)) => {
-            asm.push(format!("<{}>", prop));
-            asm.push(format!("<{}>", prop2));
-            asm.push(OP_EQUAL.to_string());
-        }
-        (Expression::Property(prop), ">=", Expression::Property(prop2)) => {
-            asm.push(format!("<{}>", prop));
-            asm.push(format!("<{}>", prop2));
-            asm.push(OP_GREATERTHANOREQUAL.to_string());
-        }
-        (Expression::CurrentInput(property), "==", Expression::Literal(value)) => {
-            if value == "true" {
-                if let Some(prop) = property {
-                    match prop.as_str() {
-                        "scriptPubKey" => asm.push(OP_INPUTBYTECODE.to_string()),
-                        "value" => asm.push(OP_INPUTVALUE.to_string()),
-                        "sequence" => asm.push(OP_INPUTSEQUENCE.to_string()),
-                        "outpoint" => asm.push(OP_INPUTOUTPOINT.to_string()),
-                        _ => asm.push(OP_INPUTBYTECODE.to_string()),
-                    }
-                } else {
-                    asm.push(OP_INPUTBYTECODE.to_string());
-                }
-            }
-        }
-        _ => {
-            // For all other expression types, delegate to emit_comparison_asm
-            emit_comparison_asm(left, op, right, asm);
-        }
-    }
-}
-
-/// Emit assembly for a comparison requirement.
-///
-/// Handles both simple comparisons (variable/literal/property) and complex
-/// expressions involving asset lookups and 64-bit arithmetic.
-fn emit_comparison_asm(left: &Expression, op: &str, right: &Expression, asm: &mut Vec<String>) {
-    // Special case: standalone property/function call introspection (dummy comparison)
-    if op == "==" {
-        if let Expression::Literal(val) = right {
-            if val == "true" {
-                // Bare `tx.assetGroups.find(...)`: the find already asserts
-                // existence via its internal OP_VERIFY and leaves the resolved
-                // packet position k. k is NOT a boolean (k == 0 is a valid
-                // successful find), so drop it and leave an explicit OP_1 as the
-                // requirement's true result.
-                if let Expression::GroupFind {
-                    asset_txid,
-                    asset_gidx,
-                } = left
-                {
-                    emit_group_find_asm(asset_txid, asset_gidx, asm);
-                    asm.push(OP_DROP.to_string());
-                    asm.push(OP_1.to_string());
-                    return;
-                }
-                // Other dummy-wrapped expressions (has/controlIs/checkSig/…)
-                // already leave a boolean — emit directly.
-                emit_expression_asm(left, asm);
-                return;
-            }
-        }
-    }
-
-    // Determine if this comparison involves 64-bit values (asset lookups, group sums)
-    let is_64bit = is_64bit_expression(left) || is_64bit_expression(right);
-
-    // Emit left operand
-    emit_expression_asm(left, asm);
-
-    // Emit right operand
-    emit_expression_asm(right, asm);
-
-    // Emit comparison operator (correct Bitcoin Script order: left, right, op)
-    if is_64bit {
-        emit_comparison_op_64(op, asm);
-    } else {
-        emit_comparison_op(op, asm);
-    }
-}
-
-/// Check if an expression produces a 64-bit (u64le) value
-fn is_64bit_expression(expr: &Expression) -> bool {
-    match expr {
-        Expression::AssetLookup { .. } => true,
-        Expression::GroupSum { .. } => true,
-        // AssetAt with "amount" property returns u64
-        Expression::AssetAt { property, .. } => property == "amount",
-        // Input/Output "value" property returns u64
-        Expression::InputIntrospection { property, .. } => property == "value",
-        Expression::OutputIntrospection { property, .. } => property == "value",
-        Expression::BinaryOp { left, right, .. } => {
-            is_64bit_expression(left) || is_64bit_expression(right)
-        }
-        _ => false,
-    }
-}
-
-/// Emit assembly for an expression (push its value onto the stack)
-fn emit_expression_asm(expr: &Expression, asm: &mut Vec<String>) {
-    match expr {
-        Expression::Variable(var) => {
-            asm.push(format!("<{}>", var));
-        }
-        Expression::Literal(lit) => {
-            asm.push(lit.clone());
-        }
-        Expression::Property(prop) => {
-            // Map the introspector "this" properties to their dedicated opcodes
-            // (the parser stores them as Property strings; resolving them here
-            // keeps the placeholder pipeline untouched for everything else).
-            match prop.as_str() {
-                "this.activeInputIndex" => asm.push(OP_PUSHCURRENTINPUTINDEX.to_string()),
-                "this.activeBytecode" => asm.push(OP_INPUTBYTECODE.to_string()),
-                _ => asm.push(format!("<{}>", prop)),
-            }
-        }
-        Expression::CurrentInput(property) => {
-            emit_current_input_asm(property.as_deref(), asm);
-        }
-        Expression::AssetLookup {
-            source,
-            index,
-            asset_txid,
-            asset_gidx,
-        } => {
-            emit_asset_lookup_asm(source, index, asset_txid, asset_gidx, asm);
-        }
-        Expression::AssetHas {
-            source,
-            index,
-            asset_txid,
-            asset_gidx,
-        } => {
-            emit_asset_has_asm(source, index, asset_txid, asset_gidx, asm);
-        }
-        Expression::AssetCount { source, index } => {
-            emit_asset_count_asm(source, index, asm);
-        }
-        Expression::AssetAt {
-            source,
-            io_index,
-            asset_index,
-            property,
-        } => {
-            emit_asset_at_asm(source, io_index, asset_index, property, asm);
-        }
-        Expression::TxIntrospection { property } => {
-            emit_tx_introspection_asm(property, asm);
-        }
-        Expression::InputIntrospection { index, property } => {
-            emit_input_introspection_asm(index, property, asm);
-        }
-        Expression::OutputIntrospection { index, property } => {
-            emit_output_introspection_asm(index, property, asm);
-        }
-        Expression::BinaryOp { left, op, right } => {
-            emit_binary_op_asm(left, op, right, asm);
-        }
-        Expression::GroupFind {
-            asset_txid,
-            asset_gidx,
-        } => {
-            emit_group_find_asm(asset_txid, asset_gidx, asm);
-        }
-        Expression::GroupHas {
-            asset_txid,
-            asset_gidx,
-        } => {
-            emit_group_has_asm(asset_txid, asset_gidx, asm);
-        }
-        Expression::GroupControlIs {
-            group,
-            asset_txid,
-            asset_gidx,
-        } => {
-            emit_group_control_is_asm(group, asset_txid, asset_gidx, asm);
-        }
-        Expression::GroupProperty { group, property } => {
-            emit_group_property_asm(group, property, asm);
-        }
-        Expression::AssetGroupsLength => {
-            asm.push(OP_INSPECTNUMASSETGROUPS.to_string());
-        }
-        Expression::GroupSum { index, source } => {
-            emit_expression_asm(index, asm);
-            match source {
-                GroupSumSource::Inputs => asm.push(OP_0.to_string()),
-                GroupSumSource::Outputs => asm.push(OP_1.to_string()),
-            }
-            asm.push(OP_INSPECTASSETGROUPSUM.to_string());
-        }
-        Expression::GroupNumIO { index, source } => {
-            emit_expression_asm(index, asm);
-            match source {
-                GroupIOSource::Inputs => asm.push(OP_0.to_string()),
-                GroupIOSource::Outputs => asm.push(OP_1.to_string()),
-            }
-            asm.push(OP_INSPECTASSETGROUPNUM.to_string());
-        }
-        Expression::GroupIOAccess {
-            group_index,
-            io_index,
-            source,
-            property,
-        } => {
-            emit_expression_asm(group_index, asm);
-            emit_expression_asm(io_index, asm);
-            match source {
-                GroupIOSource::Inputs => asm.push(OP_0.to_string()),
-                GroupIOSource::Outputs => asm.push(OP_1.to_string()),
-            }
-            asm.push(OP_INSPECTASSETGROUP.to_string());
-            // Extract property if specified
-            if let Some(prop) = property {
-                match prop.as_str() {
-                    "amount" => {
-                        // Amount is on top, no extraction needed for amount
-                    }
-                    "type" => {
-                        asm.push(OP_DROP.to_string()); // amount
-                        asm.push(OP_DROP.to_string()); // data
-                    }
-                    _ => {}
-                }
-            }
-        }
-        Expression::ContractInstance {
-            contract_name,
-            args,
-        } => {
-            emit_contract_instance_asm(contract_name, args, asm);
-        }
-        Expression::CheckSigExpr { signature, pubkey } => {
-            asm.push(format!("<{}>", pubkey));
-            asm.push(format!("<{}>", signature));
-            asm.push(OP_CHECKSIG.to_string());
-        }
-        Expression::CheckSigFromStackExpr {
-            signature,
-            pubkey,
-            message,
-        } => {
-            asm.push(format!("<{}>", message));
-            asm.push(format!("<{}>", pubkey));
-            asm.push(format!("<{}>", signature));
-            asm.push(OP_CHECKSIGFROMSTACK.to_string());
-        }
-        // Streaming SHA256
-        Expression::Sha256 { data } => {
-            emit_expression_asm(data, asm);
-            asm.push(OP_SHA256.to_string());
-        }
-        Expression::Sha256Initialize { data } => {
-            emit_expression_asm(data, asm);
-            asm.push(OP_SHA256INITIALIZE.to_string());
-        }
-        Expression::Sha256Update { context, chunk } => {
-            emit_expression_asm(context, asm);
-            emit_expression_asm(chunk, asm);
-            asm.push(OP_SHA256UPDATE.to_string());
-        }
-        Expression::Sha256Finalize {
-            context,
-            last_chunk,
-        } => {
-            emit_expression_asm(context, asm);
-            emit_expression_asm(last_chunk, asm);
-            asm.push(OP_SHA256FINALIZE.to_string());
-        }
-        // Byte-string concatenation: bytes-like + value → OP_CAT
-        Expression::Concat {
-            left,
-            right,
-            coerce_left,
-            coerce_right,
-        } => {
-            emit_expression_asm(left, asm);
-            if *coerce_left {
-                asm.push(OP_SCRIPTNUMTOLE64.to_string());
-            }
-            emit_expression_asm(right, asm);
-            if *coerce_right {
-                asm.push(OP_SCRIPTNUMTOLE64.to_string());
-            }
-            asm.push(OP_CAT.to_string());
-        }
-        // Conversion & Arithmetic
-        Expression::Neg64 { value } => {
-            emit_expression_asm(value, asm);
-            asm.push(OP_NEG64.to_string());
-        }
-        Expression::Le64ToScriptNum { value } => {
-            emit_expression_asm(value, asm);
-            asm.push(OP_LE64TOSCRIPTNUM.to_string());
-        }
-        Expression::Le32ToLe64 { value } => {
-            emit_expression_asm(value, asm);
-            asm.push(OP_LE32TOLE64.to_string());
-        }
-        // Crypto Opcodes
-        Expression::EcMulScalarVerify {
-            scalar,
-            point_p,
-            point_q,
-        } => {
-            emit_expression_asm(point_q, asm);
-            emit_expression_asm(point_p, asm);
-            emit_expression_asm(scalar, asm);
-            asm.push(OP_ECMULSCALARVERIFY.to_string());
-        }
-        Expression::TweakVerify {
-            point_p,
-            tweak,
-            point_q,
-        } => {
-            emit_expression_asm(point_q, asm);
-            emit_expression_asm(tweak, asm);
-            emit_expression_asm(point_p, asm);
-            asm.push(OP_TWEAKVERIFY.to_string());
-        }
-        Expression::CheckSigFromStackVerify {
-            signature,
-            pubkey,
-            message,
-        } => {
-            asm.push(format!("<{}>", message));
-            asm.push(format!("<{}>", pubkey));
-            asm.push(format!("<{}>", signature));
-            asm.push(OP_CHECKSIGFROMSTACKVERIFY.to_string());
-        }
-        // Byte-string manipulation (introspector extensions)
-        Expression::Substr { data, offset, size } => {
-            emit_expression_asm(data, asm);
-            emit_expression_asm(offset, asm);
-            emit_expression_asm(size, asm);
-            asm.push(OP_SUBSTR.to_string());
-        }
-        Expression::Cat { left, right } => {
-            emit_expression_asm(left, asm);
-            emit_expression_asm(right, asm);
-            asm.push(OP_CAT.to_string());
-        }
-        Expression::Bin2Num { data } => {
-            emit_expression_asm(data, asm);
-            asm.push(OP_BIN2NUM.to_string());
-        }
-        Expression::Num2Bin { value, size } => {
-            emit_expression_asm(value, asm);
-            emit_expression_asm(size, asm);
-            asm.push(OP_NUM2BIN.to_string());
-        }
-        Expression::SizeOf { data } => {
-            emit_expression_asm(data, asm);
-            asm.push(OP_SIZE.to_string());
-            asm.push(OP_NIP.to_string());
-        }
-        // Packet introspection
-        Expression::PacketInspect { packet_type } => {
-            emit_expression_asm(packet_type, asm);
-            asm.push(OP_INSPECTPACKET.to_string());
-            asm.push(OP_1.to_string());
-            asm.push(OP_EQUALVERIFY.to_string());
-        }
-        Expression::InputPacketInspect { index, packet_type } => {
-            emit_expression_asm(packet_type, asm);
-            emit_expression_asm(index, asm);
-            asm.push(OP_INSPECTINPUTPACKET.to_string());
-            asm.push(OP_1.to_string());
-            asm.push(OP_EQUALVERIFY.to_string());
-        }
-    }
-}
-
-/// Emit assembly for tx.input.current property access
-fn emit_current_input_asm(property: Option<&str>, asm: &mut Vec<String>) {
-    match property {
-        Some("scriptPubKey") => {
-            asm.push(OP_PUSHCURRENTINPUTINDEX.to_string());
-            asm.push(OP_INSPECTINPUTSCRIPTPUBKEY.to_string());
-        }
-        Some("value") => {
-            asm.push(OP_PUSHCURRENTINPUTINDEX.to_string());
-            asm.push(OP_INSPECTINPUTVALUE.to_string());
-        }
-        Some("sequence") => {
-            asm.push(OP_PUSHCURRENTINPUTINDEX.to_string());
-            asm.push(OP_INSPECTINPUTSEQUENCE.to_string());
-        }
-        Some("outpoint") => {
-            asm.push(OP_PUSHCURRENTINPUTINDEX.to_string());
-            asm.push(OP_INSPECTINPUTOUTPOINT.to_string());
-        }
-        _ => {
-            asm.push(OP_PUSHCURRENTINPUTINDEX.to_string());
-            asm.push(OP_INSPECTINPUTSCRIPTPUBKEY.to_string());
-        }
-    }
-}
-
-/// Emit assembly for a contract instantiation: `new ContractName(arg1, arg2, ...)`
-///
-/// Produces a single placeholder token `<VTXO:ContractName(<arg1>,<arg2>)>` that
-/// the runtime resolves to the Taproot scriptPubKey (34-byte P2TR output script)
-/// of the named contract instantiated with the given constructor arguments.
-///
-/// Options (server key, exit timelock) are inherited from the enclosing contract
-/// and must be applied by the runtime when computing the child contract's taproot
-/// key.
-///
-/// Typical usage (recursion / self-referential contract enforcement):
-/// ```text
-/// require(tx.outputs[0].scriptPubKey == new SingleSig(ownerPk));
-/// ```
-/// compiles to:
-/// ```text
-/// 0 OP_INSPECTOUTPUTSCRIPTPUBKEY <VTXO:SingleSig(<ownerPk>)> OP_EQUAL
-/// ```
-fn emit_contract_instance_asm(contract_name: &str, args: &[Expression], asm: &mut Vec<String>) {
-    let args_str = args
-        .iter()
-        .map(|a| match a {
-            Expression::Variable(v) => format!("<{}>", v),
-            Expression::Literal(l) => l.clone(),
-            _ => {
-                // For complex arg expressions, emit a nested representation
-                let mut nested = Vec::new();
-                emit_expression_asm(a, &mut nested);
-                nested.join(" ")
-            }
-        })
-        .collect::<Vec<_>>()
-        .join(",");
-
-    asm.push(format!("<VTXO:{}({})>", contract_name, args_str));
-}
-
-/// Push the lookup operands and the source-appropriate lookup opcode.
-///
-/// Push order follows the opcode (`popAssetID` pops gidx top, then txid; then
-/// the io index `o`/`i`): index, then txid, then gidx. Stack after the opcode is
-/// `[amount, success_flag]` (flag on top).
-fn emit_asset_lookup_operands(
-    source: &AssetLookupSource,
-    index: &Expression,
-    asset_txid: &Expression,
-    asset_gidx: &Expression,
-    asm: &mut Vec<String>,
-) {
-    emit_expression_asm(index, asm);
-    emit_expression_asm(asset_txid, asm); // -> <fooTxid>
-    emit_expression_asm(asset_gidx, asm); // -> <fooGidx> or pushed literal
-    match source {
-        AssetLookupSource::Input => asm.push(OP_INSPECTINASSETLOOKUP.to_string()),
-        AssetLookupSource::Output => asm.push(OP_INSPECTOUTASSETLOOKUP.to_string()),
-    }
-}
-
-/// Emit `tx.{inputs,outputs}[i].assets.lookup(txid, gidx)`: assert the asset is
-/// present (consume the success flag with OP_VERIFY) and leave the typed amount.
-fn emit_asset_lookup_asm(
-    source: &AssetLookupSource,
-    index: &Expression,
-    asset_txid: &Expression,
-    asset_gidx: &Expression,
-    asm: &mut Vec<String>,
-) {
-    emit_asset_lookup_operands(source, index, asset_txid, asset_gidx, asm);
-    asm.push(OP_VERIFY.to_string()); // consume success flag, leave amount
-}
-
-/// Emit `tx.{inputs,outputs}[i].assets.has(txid, gidx)`: boolean presence —
-/// keep the success flag, drop the amount below it with OP_NIP.
-fn emit_asset_has_asm(
-    source: &AssetLookupSource,
-    index: &Expression,
-    asset_txid: &Expression,
-    asset_gidx: &Expression,
-    asm: &mut Vec<String>,
-) {
-    emit_asset_lookup_operands(source, index, asset_txid, asset_gidx, asm);
-    asm.push(OP_NIP.to_string()); // drop amount, leave success flag (Bool)
-}
-
-/// Emit `tx.assetGroups.find(txid, gidx)`: assert existence (consume the success
-/// flag with OP_VERIFY) and leave the resolved packet position k.
-fn emit_group_find_asm(asset_txid: &Expression, asset_gidx: &Expression, asm: &mut Vec<String>) {
-    emit_expression_asm(asset_txid, asm);
-    emit_expression_asm(asset_gidx, asm);
-    asm.push(OP_FINDASSETGROUPBYASSETID.to_string());
-    asm.push(OP_VERIFY.to_string()); // consume success flag, leave k; fail if absent
-}
-
-/// Emit `tx.assetGroups.has(txid, gidx)`: boolean presence — keep the success
-/// flag, drop the resolved position k below it with OP_NIP.
-fn emit_group_has_asm(asset_txid: &Expression, asset_gidx: &Expression, asm: &mut Vec<String>) {
-    emit_expression_asm(asset_txid, asm);
-    emit_expression_asm(asset_gidx, asm);
-    asm.push(OP_FINDASSETGROUPBYASSETID.to_string());
-    asm.push(OP_NIP.to_string()); // drop k, leave success flag (Bool)
-}
-
-/// Emit `group.controlIs(txid, gidx)`: boolean equality over the complete
-/// canonical control Asset ID. Stack after OP_INSPECTASSETGROUPCTRL is
-/// `[ctrl_txid, ctrl_gidx, flag]`; drop the flag, compare gidx then txid (no
-/// OP_EQUALVERIFY), and AND the two booleans. The absent tuple
-/// `[empty_bytes, 0, 0]` cannot equal a valid bytes32 txid, so absence is false.
-fn emit_group_control_is_asm(
-    group: &str,
-    asset_txid: &Expression,
-    asset_gidx: &Expression,
-    asm: &mut Vec<String>,
-) {
-    asm.push(format!("<{}>", group));
-    asm.push(OP_INSPECTASSETGROUPCTRL.to_string());
-    asm.push(OP_DROP.to_string()); // drop success flag -> [ctrl_txid, ctrl_gidx]
-    emit_expression_asm(asset_gidx, asm);
-    asm.push(OP_EQUAL.to_string()); // ctrl_gidx == gidx -> [ctrl_txid, bool]
-    asm.push(OP_SWAP.to_string()); // -> [bool, ctrl_txid]
-    emit_expression_asm(asset_txid, asm);
-    asm.push(OP_EQUAL.to_string()); // ctrl_txid == txid -> [bool, bool]
-    asm.push(OP_BOOLAND.to_string()); // combine -> Bool
-}
-
-/// Emit assembly for asset count: tx.inputs[i].assets.length or tx.outputs[o].assets.length
-///
-/// Pushes the count of assets at the given input/output index.
-fn emit_asset_count_asm(source: &AssetLookupSource, index: &Expression, asm: &mut Vec<String>) {
-    // Push the index
-    emit_expression_asm(index, asm);
-
-    // Emit the appropriate count opcode
-    match source {
-        AssetLookupSource::Input => {
-            asm.push(OP_INSPECTINASSETCOUNT.to_string());
-        }
-        AssetLookupSource::Output => {
-            asm.push(OP_INSPECTOUTASSETCOUNT.to_string());
-        }
-    }
-}
-
-/// Emit assembly for indexed asset access: tx.inputs[i].assets[t].property
-///
-/// OP_INSPECTINASSETAT / OP_INSPECTOUTASSETAT returns: txid32, gidx_u16, amount_u64
-/// We extract based on the property requested.
-fn emit_asset_at_asm(
-    source: &AssetLookupSource,
-    io_index: &Expression,
-    asset_index: &Expression,
-    property: &str,
-    asm: &mut Vec<String>,
-) {
-    // Push io_index
-    emit_expression_asm(io_index, asm);
-
-    // Push asset_index
-    emit_expression_asm(asset_index, asm);
-
-    // Emit the appropriate opcode
-    match source {
-        AssetLookupSource::Input => {
-            asm.push(OP_INSPECTINASSETAT.to_string());
-        }
-        AssetLookupSource::Output => {
-            asm.push(OP_INSPECTOUTASSETAT.to_string());
-        }
-    }
-
-    // Stack after opcode: txid32, gidx_u16, amount_u64 (top)
-    // Extract based on property
-    match property {
-        "assetId" => {
-            // Drop the amount, keep the canonical Asset ID (asset_txid, asset_gidx).
-            // TODO(asset-id-struct): this intentionally leaves TWO stack items, so
-            // `.assetId` needs a composite `AssetId` struct return type before it
-            // can be destructured (.txid/.gidx) or compared safely. Deferred to a
-            // separate PR.
-            asm.push(OP_DROP.to_string());
-        }
-        "amount" => {
-            // Keep only the amount (top of stack)
-            // NIP removes the second item from the top
-            asm.push(OP_NIP.to_string()); // Remove gidx_u16
-            asm.push(OP_NIP.to_string()); // Remove txid32
-        }
-        _ => {
-            // Unknown property, leave stack as-is
-        }
-    }
-}
-
-/// Emit assembly for transaction introspection: tx.version, tx.locktime, etc.
-fn emit_tx_introspection_asm(property: &str, asm: &mut Vec<String>) {
-    match property {
-        "version" => asm.push(OP_INSPECTVERSION.to_string()),
-        "locktime" => asm.push(OP_INSPECTLOCKTIME.to_string()),
-        "numInputs" => asm.push(OP_INSPECTNUMINPUTS.to_string()),
-        "numOutputs" => asm.push(OP_INSPECTNUMOUTPUTS.to_string()),
-        "weight" => asm.push(OP_TXWEIGHT.to_string()),
-        "id" => asm.push(OP_TXID.to_string()),
-        _ => {
-            // Unknown property, emit as placeholder
-            asm.push(format!("<tx.{}>", property));
-        }
-    }
-}
-
-/// Emit assembly for input introspection: tx.inputs[i].property
-fn emit_input_introspection_asm(index: &Expression, property: &str, asm: &mut Vec<String>) {
-    // Push the index
-    emit_expression_asm(index, asm);
-
-    // Emit the appropriate opcode
-    match property {
-        "value" => asm.push(OP_INSPECTINPUTVALUE.to_string()),
-        "scriptPubKey" => asm.push(OP_INSPECTINPUTSCRIPTPUBKEY.to_string()),
-        "sequence" => asm.push(OP_INSPECTINPUTSEQUENCE.to_string()),
-        "outpoint" => asm.push(OP_INSPECTINPUTOUTPOINT.to_string()),
-        "issuance" => asm.push(OP_INSPECTINPUTISSUANCE.to_string()),
-        "arkadeScriptHash" => asm.push(OP_INSPECTINPUTARKADESCRIPTHASH.to_string()),
-        "arkadeWitnessHash" => asm.push(OP_INSPECTINPUTARKADEWITNESSHASH.to_string()),
-        _ => {
-            // Unknown property, emit as placeholder
-            asm.push(format!("<tx.inputs[?].{}>", property));
-        }
-    }
-}
-
-/// Emit assembly for output introspection: tx.outputs[o].property
-fn emit_output_introspection_asm(index: &Expression, property: &str, asm: &mut Vec<String>) {
-    // Push the index
-    emit_expression_asm(index, asm);
-
-    // Emit the appropriate opcode
-    match property {
-        "value" => asm.push(OP_INSPECTOUTPUTVALUE.to_string()),
-        "scriptPubKey" => asm.push(OP_INSPECTOUTPUTSCRIPTPUBKEY.to_string()),
-        "nonce" => asm.push(OP_INSPECTOUTPUTNONCE.to_string()),
-        _ => {
-            // Unknown property, emit as placeholder
-            asm.push(format!("<tx.outputs[?].{}>", property));
-        }
-    }
-}
-
-/// Emit assembly for a binary arithmetic operation (64-bit)
-fn emit_binary_op_asm(left: &Expression, op: &str, right: &Expression, asm: &mut Vec<String>) {
-    // Emit left operand
-    emit_expression_asm(left, asm);
-
-    // Convert to u64le if needed (witness inputs arrive as csn)
-    if needs_u64_conversion(left) {
-        asm.push(OP_SCRIPTNUMTOLE64.to_string());
-    }
-
-    // Emit right operand
-    emit_expression_asm(right, asm);
-
-    // Convert to u64le if needed
-    if needs_u64_conversion(right) {
-        asm.push(OP_SCRIPTNUMTOLE64.to_string());
-    }
-
-    // Emit 64-bit arithmetic opcode + overflow verify
-    match op {
-        "+" => {
-            asm.push(OP_ADD64.to_string());
-            asm.push(OP_VERIFY.to_string());
-        }
-        "-" => {
-            asm.push(OP_SUB64.to_string());
-            asm.push(OP_VERIFY.to_string());
-        }
-        "*" => {
-            asm.push(OP_MUL64.to_string());
-            asm.push(OP_VERIFY.to_string());
-        }
-        "/" => {
-            asm.push(OP_DIV64.to_string());
-            asm.push(OP_VERIFY.to_string());
-        }
-        _ => {
-            asm.push(format!("OP_{}", op.to_uppercase()));
-        }
-    }
-}
-
-/// Check if an expression needs csn→u64le conversion for 64-bit arithmetic
-fn needs_u64_conversion(expr: &Expression) -> bool {
-    match expr {
-        // Variables (witness inputs) arrive as CScriptNum
-        Expression::Variable(_) => true,
-        // Literals are emitted as-is (caller should provide 8-byte LE)
-        Expression::Literal(_) => false,
-        // Asset lookups already produce u64le
-        Expression::AssetLookup { .. } => false,
-        // Asset count produces CScriptNum
-        Expression::AssetCount { .. } => false,
-        // AssetAt "amount" produces u64le, "assetId" produces (txid32, gidx_u16)
-        Expression::AssetAt { property, .. } => property != "amount",
-        // Group sums already produce u64le
-        Expression::GroupSum { .. } => false,
-        // Binary ops produce u64le
-        Expression::BinaryOp { .. } => false,
-        // Properties depend on context
-        Expression::Property(_) => false,
-        _ => false,
-    }
-}
-
-/// Emit assembly for group property access
-fn emit_group_property_asm(group: &str, property: &str, asm: &mut Vec<String>) {
-    match property {
-        "sumInputs" => {
-            asm.push(format!("<{}>", group));
-            asm.push(OP_0.to_string()); // source=inputs
-            asm.push(OP_INSPECTASSETGROUPSUM.to_string());
-        }
-        "sumOutputs" => {
-            asm.push(format!("<{}>", group));
-            asm.push(OP_1.to_string()); // source=outputs
-            asm.push(OP_INSPECTASSETGROUPSUM.to_string());
-        }
-        "numInputs" => {
-            asm.push(format!("<{}>", group));
-            asm.push(OP_0.to_string()); // source=inputs
-            asm.push(OP_INSPECTASSETGROUPNUM.to_string());
-        }
-        "numOutputs" => {
-            asm.push(format!("<{}>", group));
-            asm.push(OP_1.to_string()); // source=outputs
-            asm.push(OP_INSPECTASSETGROUPNUM.to_string());
-        }
-        "delta" => {
-            // delta = sumOutputs - sumInputs
-            asm.push(format!("<{}>", group));
-            asm.push(OP_1.to_string());
-            asm.push(OP_INSPECTASSETGROUPSUM.to_string());
-            asm.push(format!("<{}>", group));
-            asm.push(OP_0.to_string());
-            asm.push(OP_INSPECTASSETGROUPSUM.to_string());
-            asm.push(OP_SUB64.to_string());
-            asm.push(OP_VERIFY.to_string());
-        }
-        "hasControl" => {
-            // group.hasControl: presence only.
-            // [ctrl_txid, ctrl_gidx, flag] -> OP_NIP OP_NIP -> flag (Bool)
-            asm.push(format!("<{}>", group));
-            asm.push(OP_INSPECTASSETGROUPCTRL.to_string());
-            asm.push(OP_NIP.to_string());
-            asm.push(OP_NIP.to_string());
-        }
-        "metadataHash" => {
-            asm.push(format!("<{}>", group));
-            asm.push(OP_INSPECTASSETGROUPMETADATAHASH.to_string());
-        }
-        "assetId" => {
-            // Returns the canonical Asset ID (asset_txid, asset_gidx) — TWO stack items.
-            // TODO(asset-id-struct): like asset_at `.assetId`, this needs a composite
-            // `AssetId` struct return type before it can be destructured (.txid/.gidx)
-            // or compared with `==` (a single OP_EQUAL only sees the top item, the
-            // gidx). Deferred to a separate PR.
-            asm.push(format!("<{}>", group));
-            asm.push(OP_INSPECTASSETGROUPASSETID.to_string());
-        }
-        "isFresh" => {
-            // isFresh: compares assetId.txid with current transaction's txid
-            // 1. Get group's assetId (returns txid32, gidx_u16)
-            asm.push(format!("<{}>", group));
-            asm.push(OP_INSPECTASSETGROUPASSETID.to_string());
-            // 2. Drop gidx_u16, keep txid32
-            asm.push(OP_DROP.to_string());
-            // 3. Get current transaction hash
-            asm.push(OP_TXHASH.to_string());
-            // 4. Compare txids - result is bool
-            asm.push(OP_EQUAL.to_string());
-        }
-        _ => {
-            // Unknown group property
-            asm.push(format!("<{}.{}>", group, property));
-        }
-    }
-}
-
-/// Emit standard comparison operator (CScriptNum / non-64-bit)
-fn emit_comparison_op(op: &str, asm: &mut Vec<String>) {
-    match op {
-        "==" => asm.push(OP_EQUAL.to_string()),
-        "!=" => {
-            asm.push(OP_EQUAL.to_string());
-            asm.push(OP_NOT.to_string());
-        }
-        ">=" => asm.push(OP_GREATERTHANOREQUAL.to_string()),
-        ">" => asm.push(OP_GREATERTHAN.to_string()),
-        "<=" => asm.push(OP_LESSTHANOREQUAL.to_string()),
-        "<" => asm.push(OP_LESSTHAN.to_string()),
-        _ => asm.push(format!("OP_{}", op)),
-    }
-}
-
-/// Emit 64-bit comparison operator (u64le operands)
-fn emit_comparison_op_64(op: &str, asm: &mut Vec<String>) {
-    match op {
-        "==" => {
-            asm.push(OP_EQUAL.to_string());
-            asm.push(OP_VERIFY.to_string());
-        }
-        "!=" => {
-            asm.push(OP_EQUAL.to_string());
-            asm.push(OP_NOT.to_string());
-            asm.push(OP_VERIFY.to_string());
-        }
-        ">=" => {
-            asm.push(OP_GREATERTHANOREQUAL64.to_string());
-            asm.push(OP_VERIFY.to_string());
-        }
-        ">" => {
-            asm.push(OP_GREATERTHAN64.to_string());
-            asm.push(OP_VERIFY.to_string());
-        }
-        "<=" => {
-            asm.push(OP_LESSTHANOREQUAL64.to_string());
-            asm.push(OP_VERIFY.to_string());
-        }
-        "<" => {
-            asm.push(OP_LESSTHAN64.to_string());
-            asm.push(OP_VERIFY.to_string());
-        }
-        _ => {
-            asm.push(format!("OP_{}", op));
-            asm.push(OP_VERIFY.to_string());
-        }
-    }
-}
-
-// ─── Loop Unrolling ─────────────────────────────────────────────────────────────
-
-fn unroll_loop_body(
-    body: &[Statement],
-    index_var: &str,
-    value_var: &str,
-    array_name: Option<&str>,
-    asm: &mut Vec<String>,
-) -> Result<(), String> {
-    for k in 0..DEFAULT_ARRAY_LENGTH {
-        let substituted_body = substitute_loop_body(body, index_var, value_var, k, array_name);
-        generate_asm_from_statements_recursive(&substituted_body, asm)?;
-    }
-    Ok(())
-}
-
-/// Substitute loop variables in the body for a specific iteration index k.
-///
-/// Transforms:
-/// - `GroupProperty { group: value_var, property: "sumOutputs" }` → `GroupSum { index: k, source: Outputs }`
-/// - `GroupProperty { group: value_var, property: "sumInputs" }` → `GroupSum { index: k, source: Inputs }`
-/// - `Variable(index_var)` → `Literal(k)`
-/// - `Variable(value_var)` when array_name is Some → `Variable("array_name_{k}")`
-/// - Property-form indexing `arr[index_var]` → `Variable("arr_{k}")`
-fn substitute_loop_body(
-    body: &[Statement],
-    index_var: &str,
-    value_var: &str,
-    k: usize,
-    array_name: Option<&str>,
-) -> Vec<Statement> {
-    body.iter()
-        .map(|stmt| substitute_statement(stmt, index_var, value_var, k, array_name))
-        .collect()
-}
-
-fn substitute_statement(
-    stmt: &Statement,
-    index_var: &str,
-    value_var: &str,
-    k: usize,
-    array_name: Option<&str>,
-) -> Statement {
-    match stmt {
-        Statement::Require(req) => Statement::Require(substitute_requirement(
-            req, index_var, value_var, k, array_name,
-        )),
-        Statement::LetBinding { name, value } => Statement::LetBinding {
-            name: name.clone(),
-            value: substitute_expression(value, index_var, value_var, k, array_name),
-        },
-        Statement::VarAssign { name, value } => Statement::VarAssign {
-            name: name.clone(),
-            value: substitute_expression(value, index_var, value_var, k, array_name),
-        },
-        Statement::IfElse {
-            condition,
-            then_body,
-            else_body,
-        } => Statement::IfElse {
-            condition: substitute_expression(condition, index_var, value_var, k, array_name),
-            then_body: substitute_loop_body(then_body, index_var, value_var, k, array_name),
-            else_body: else_body
-                .as_ref()
-                .map(|b| substitute_loop_body(b, index_var, value_var, k, array_name)),
-        },
-        Statement::ForIn {
-            index_var: inner_idx,
-            value_var: inner_val,
-            iterable,
-            body,
-        } => {
-            // Nested loops: substitute in iterable, leave inner variables alone
-            Statement::ForIn {
-                index_var: inner_idx.clone(),
-                value_var: inner_val.clone(),
-                iterable: substitute_expression(iterable, index_var, value_var, k, array_name),
-                body: body.clone(), // Inner loop body keeps its own variables
-            }
-        }
-    }
-}
-
-fn substitute_requirement(
-    req: &Requirement,
-    index_var: &str,
-    value_var: &str,
-    k: usize,
-    array_name: Option<&str>,
-) -> Requirement {
-    match req {
-        Requirement::Comparison { left, op, right } => Requirement::Comparison {
-            left: substitute_expression(left, index_var, value_var, k, array_name),
-            op: op.clone(),
-            right: substitute_expression(right, index_var, value_var, k, array_name),
-        },
-        Requirement::CheckSig { signature, pubkey } => {
-            // Substitute signature and pubkey if they match loop variables
-            let new_sig = if signature == value_var {
-                if let Some(arr) = array_name {
-                    format!("{}_{}", arr, k)
-                } else {
-                    signature.clone()
-                }
-            } else {
-                signature.clone()
-            };
-            Requirement::CheckSig {
-                signature: new_sig,
-                pubkey: pubkey.clone(),
-            }
-        }
-        Requirement::CheckSigFromStack {
-            signature,
-            pubkey,
-            message,
-        } => {
-            // Substitute signature, pubkey, and message if they match loop variables
-            let new_sig = if signature == value_var {
-                if let Some(arr) = array_name {
-                    format!("{}_{}", arr, k)
-                } else {
-                    signature.clone()
-                }
-            } else {
-                signature.clone()
-            };
-            Requirement::CheckSigFromStack {
-                signature: new_sig,
-                pubkey: pubkey.clone(),
-                message: message.clone(),
-            }
-        }
-        // Other requirement types don't need substitution
-        _ => req.clone(),
-    }
-}
-
-fn substitute_expression(
-    expr: &Expression,
-    index_var: &str,
-    value_var: &str,
-    k: usize,
-    array_name: Option<&str>,
-) -> Expression {
-    match expr {
-        // Replace index variable with literal k
-        Expression::Variable(var) if var == index_var => Expression::Literal(k.to_string()),
-        // Replace value_var with array_name_{k} when iterating over arrays
-        Expression::Variable(var) if var == value_var => {
-            if let Some(name) = array_name {
-                Expression::Variable(format!("{}_{}", name, k))
-            } else {
-                Expression::Variable(var.clone())
-            }
-        }
-        // Replace value_var.property with appropriate indexed expression
-        Expression::GroupProperty { group, property } if group == value_var => {
-            match property.as_str() {
-                "sumInputs" => Expression::GroupSum {
-                    index: Box::new(Expression::Literal(k.to_string())),
-                    source: GroupSumSource::Inputs,
-                },
-                "sumOutputs" => Expression::GroupSum {
-                    index: Box::new(Expression::Literal(k.to_string())),
-                    source: GroupSumSource::Outputs,
-                },
-                // For delta, control, isFresh, assetId, metadataHash - replace group name with index literal
-                _ => Expression::GroupProperty {
-                    group: k.to_string(),
-                    property: property.clone(),
-                },
-            }
-        }
-        // Handle property strings that represent array indexing (e.g., "oracles[i]").
-        Expression::Property(prop) => {
-            // Check if this looks like array indexing
-            if let Some(bracket_start) = prop.find('[') {
-                if let Some(bracket_end) = prop.find(']') {
-                    let arr_name = &prop[..bracket_start];
-                    let idx = &prop[bracket_start + 1..bracket_end];
-                    if idx == index_var {
-                        return Expression::Variable(format!("{}_{}", arr_name, k));
-                    }
-                }
-            }
-            expr.clone()
-        }
-        // Recursively substitute in binary operations
-        Expression::BinaryOp { left, op, right } => Expression::BinaryOp {
-            left: Box::new(substitute_expression(
-                left, index_var, value_var, k, array_name,
-            )),
-            op: op.clone(),
-            right: Box::new(substitute_expression(
-                right, index_var, value_var, k, array_name,
-            )),
-        },
-        // Handle CheckSigFromStackExpr
-        Expression::CheckSigFromStackExpr {
-            signature,
-            pubkey,
-            message,
-        } => {
-            let new_sig = if signature == value_var {
-                if let Some(arr) = array_name {
-                    format!("{}_{}", arr, k)
-                } else {
-                    signature.clone()
-                }
-            } else {
-                signature.clone()
-            };
-            // Check if pubkey is an array indexed expression (string form)
-            let new_pk = if pubkey.contains('[') && pubkey.contains(']') {
-                if let Some(bracket_start) = pubkey.find('[') {
-                    if let Some(bracket_end) = pubkey.find(']') {
-                        let arr_name = &pubkey[..bracket_start];
-                        let idx = &pubkey[bracket_start + 1..bracket_end];
-                        if idx == index_var {
-                            format!("{}_{}", arr_name, k)
-                        } else {
-                            pubkey.clone()
-                        }
-                    } else {
-                        pubkey.clone()
-                    }
-                } else {
-                    pubkey.clone()
-                }
-            } else {
-                pubkey.clone()
-            };
-            Expression::CheckSigFromStackExpr {
-                signature: new_sig,
-                pubkey: new_pk,
-                message: message.clone(),
-            }
-        }
-        // Handle CheckSigExpr
-        Expression::CheckSigExpr { signature, pubkey } => {
-            let new_sig = if signature == value_var {
-                if let Some(arr) = array_name {
-                    format!("{}_{}", arr, k)
-                } else {
-                    signature.clone()
-                }
-            } else {
-                signature.clone()
-            };
-            Expression::CheckSigExpr {
-                signature: new_sig,
-                pubkey: pubkey.clone(),
-            }
-        }
-        // Handle InputIntrospection - substitute index if it matches loop variable
-        Expression::InputIntrospection { index, property } => Expression::InputIntrospection {
-            index: Box::new(substitute_expression(
-                index, index_var, value_var, k, array_name,
-            )),
-            property: property.clone(),
-        },
-        // Handle OutputIntrospection - substitute index if it matches loop variable
-        Expression::OutputIntrospection { index, property } => Expression::OutputIntrospection {
-            index: Box::new(substitute_expression(
-                index, index_var, value_var, k, array_name,
-            )),
-            property: property.clone(),
-        },
-        // Recurse into contract instance arguments
-        Expression::ContractInstance {
-            contract_name,
-            args,
-        } => Expression::ContractInstance {
-            contract_name: contract_name.clone(),
-            args: args
-                .iter()
-                .map(|a| substitute_expression(a, index_var, value_var, k, array_name))
-                .collect(),
-        },
-        // Asset lookups/has: substitute the io index and both Asset ID operands
-        // (e.g. `tx.outputs[i].assets.lookup(assetTxid, i)` unrolls i -> 0,1,2…).
-        Expression::AssetLookup {
-            source,
-            index,
-            asset_txid,
-            asset_gidx,
-        } => Expression::AssetLookup {
-            source: source.clone(),
-            index: Box::new(substitute_expression(
-                index, index_var, value_var, k, array_name,
-            )),
-            asset_txid: Box::new(substitute_expression(
-                asset_txid, index_var, value_var, k, array_name,
-            )),
-            asset_gidx: Box::new(substitute_expression(
-                asset_gidx, index_var, value_var, k, array_name,
-            )),
-        },
-        Expression::AssetHas {
-            source,
-            index,
-            asset_txid,
-            asset_gidx,
-        } => Expression::AssetHas {
-            source: source.clone(),
-            index: Box::new(substitute_expression(
-                index, index_var, value_var, k, array_name,
-            )),
-            asset_txid: Box::new(substitute_expression(
-                asset_txid, index_var, value_var, k, array_name,
-            )),
-            asset_gidx: Box::new(substitute_expression(
-                asset_gidx, index_var, value_var, k, array_name,
-            )),
-        },
-        Expression::GroupFind {
-            asset_txid,
-            asset_gidx,
-        } => Expression::GroupFind {
-            asset_txid: Box::new(substitute_expression(
-                asset_txid, index_var, value_var, k, array_name,
-            )),
-            asset_gidx: Box::new(substitute_expression(
-                asset_gidx, index_var, value_var, k, array_name,
-            )),
-        },
-        Expression::GroupHas {
-            asset_txid,
-            asset_gidx,
-        } => Expression::GroupHas {
-            asset_txid: Box::new(substitute_expression(
-                asset_txid, index_var, value_var, k, array_name,
-            )),
-            asset_gidx: Box::new(substitute_expression(
-                asset_gidx, index_var, value_var, k, array_name,
-            )),
-        },
-        Expression::GroupControlIs {
-            group,
-            asset_txid,
-            asset_gidx,
-        } => Expression::GroupControlIs {
-            // Replace the group name with the loop index when iterating groups.
-            group: if group == value_var {
-                k.to_string()
-            } else {
-                group.clone()
-            },
-            asset_txid: Box::new(substitute_expression(
-                asset_txid, index_var, value_var, k, array_name,
-            )),
-            asset_gidx: Box::new(substitute_expression(
-                asset_gidx, index_var, value_var, k, array_name,
-            )),
-        },
-        // All other expressions are returned as-is
-        _ => expr.clone(),
-    }
-}
-
-// ─── Concat rewrite pass ────────────────────────────────────────────────────
-//
-// Walk every function's AST and convert `BinaryOp { op: "+" }` into
-// `Concat { ... }` when at least one operand resolves to a bytes-like type
-// (Bytes, Bytes20, Bytes32). Pure int+int additions stay as `BinaryOp` and
-// continue to emit OP_ADD64.
-//
-// We need types to make the decision, so the walk threads a `Scope` and
-// uses `typechecker::infer_type` on rewritten subtrees. The rewrite is
-// bottom-up: children are rewritten first so the parent sees the post-
-// rewrite type (e.g. `bytes32 + int` rewrites to `Concat` of type Bytes,
-// which then makes the outer `+ int` also a Concat).
-
-use crate::typechecker::{is_bytes_like, needs_scriptnum_to_le64, Scope};
-
-fn rewrite_concat_ops(contract: &mut crate::models::Contract) {
-    let constructor_scope = crate::typechecker::build_scope(&contract.parameters);
-    for function in &mut contract.functions {
-        let mut scope = constructor_scope.clone();
-        scope.extend(crate::typechecker::build_scope(&function.parameters));
-        rewrite_statements_concat(&mut function.statements, &mut scope);
-    }
-}
-
-fn rewrite_statements_concat(stmts: &mut [Statement], scope: &mut Scope) {
-    for stmt in stmts {
-        rewrite_statement_concat(stmt, scope);
-    }
-}
-
-fn rewrite_statement_concat(stmt: &mut Statement, scope: &mut Scope) {
-    match stmt {
-        Statement::Require(req) => rewrite_requirement_concat(req, scope),
-        Statement::LetBinding { name, value } => {
-            let (new_expr, t) = rewrite_expression_concat(
-                std::mem::replace(value, Expression::Literal(String::new())),
-                scope,
-            );
-            *value = new_expr;
-            scope.insert(name.clone(), t);
-        }
-        Statement::VarAssign { name, value } => {
-            let (new_expr, t) = rewrite_expression_concat(
-                std::mem::replace(value, Expression::Literal(String::new())),
-                scope,
-            );
-            *value = new_expr;
-            scope.insert(name.clone(), t);
-        }
-        Statement::IfElse {
-            condition,
-            then_body,
-            else_body,
-        } => {
-            let (new_cond, _) = rewrite_expression_concat(
-                std::mem::replace(condition, Expression::Literal(String::new())),
-                scope,
-            );
-            *condition = new_cond;
-            let mut then_scope = scope.clone();
-            rewrite_statements_concat(then_body, &mut then_scope);
-            if let Some(eb) = else_body {
-                let mut else_scope = scope.clone();
-                rewrite_statements_concat(eb, &mut else_scope);
-            }
-        }
-        Statement::ForIn {
-            index_var,
-            value_var,
-            iterable,
-            body,
-        } => {
-            let (new_iter, _) = rewrite_expression_concat(
-                std::mem::replace(iterable, Expression::Literal(String::new())),
-                scope,
-            );
-            *iterable = new_iter;
-            let mut loop_scope = scope.clone();
-            loop_scope.insert(index_var.clone(), ArkType::Int);
-            loop_scope.insert(value_var.clone(), ArkType::Unknown);
-            rewrite_statements_concat(body, &mut loop_scope);
-        }
-    }
-}
-
-fn rewrite_requirement_concat(req: &mut Requirement, scope: &Scope) {
-    if let Requirement::Comparison { left, right, .. } = req {
-        let (nl, _) = rewrite_expression_concat(
-            std::mem::replace(left, Expression::Literal(String::new())),
-            scope,
-        );
-        *left = nl;
-        let (nr, _) = rewrite_expression_concat(
-            std::mem::replace(right, Expression::Literal(String::new())),
-            scope,
-        );
-        *right = nr;
-    }
-}
-
-fn rewrite_expression_concat(expr: Expression, scope: &Scope) -> (Expression, ArkType) {
-    match expr {
-        Expression::BinaryOp { left, op, right } => {
-            let (new_l, lt) = rewrite_expression_concat(*left, scope);
-            let (new_r, rt) = rewrite_expression_concat(*right, scope);
-            if op == "+" && (is_bytes_like(&lt) || is_bytes_like(&rt)) {
-                let coerce_left = needs_scriptnum_to_le64(&lt);
-                let coerce_right = needs_scriptnum_to_le64(&rt);
-                (
-                    Expression::Concat {
-                        left: Box::new(new_l),
-                        right: Box::new(new_r),
-                        coerce_left,
-                        coerce_right,
-                    },
-                    ArkType::Bytes,
-                )
-            } else {
-                let result_type = match op.as_str() {
-                    "+" | "-" | "*" | "/" => {
-                        if lt == ArkType::Uint64Le || rt == ArkType::Uint64Le {
-                            ArkType::Uint64Le
-                        } else {
-                            ArkType::Int
-                        }
-                    }
-                    "==" | "!=" | ">=" | "<=" | ">" | "<" => ArkType::Bool,
-                    _ => ArkType::Unknown,
-                };
-                (
-                    Expression::BinaryOp {
-                        left: Box::new(new_l),
-                        op,
-                        right: Box::new(new_r),
-                    },
-                    result_type,
-                )
-            }
-        }
-        Expression::Sha256 { data } => {
-            let (new_data, _) = rewrite_expression_concat(*data, scope);
-            (
-                Expression::Sha256 {
-                    data: Box::new(new_data),
-                },
-                ArkType::Bytes32,
-            )
-        }
-        Expression::Sha256Initialize { data } => {
-            let (new_data, _) = rewrite_expression_concat(*data, scope);
-            (
-                Expression::Sha256Initialize {
-                    data: Box::new(new_data),
-                },
-                ArkType::Bytes32,
-            )
-        }
-        Expression::Sha256Update { context, chunk } => {
-            let (new_ctx, _) = rewrite_expression_concat(*context, scope);
-            let (new_chunk, _) = rewrite_expression_concat(*chunk, scope);
-            (
-                Expression::Sha256Update {
-                    context: Box::new(new_ctx),
-                    chunk: Box::new(new_chunk),
-                },
-                ArkType::Bytes32,
-            )
-        }
-        Expression::Sha256Finalize {
-            context,
-            last_chunk,
-        } => {
-            let (new_ctx, _) = rewrite_expression_concat(*context, scope);
-            let (new_chunk, _) = rewrite_expression_concat(*last_chunk, scope);
-            (
-                Expression::Sha256Finalize {
-                    context: Box::new(new_ctx),
-                    last_chunk: Box::new(new_chunk),
-                },
-                ArkType::Bytes32,
-            )
-        }
-        Expression::Concat {
-            left,
-            right,
-            coerce_left,
-            coerce_right,
-        } => {
-            let (new_l, _) = rewrite_expression_concat(*left, scope);
-            let (new_r, _) = rewrite_expression_concat(*right, scope);
-            (
-                Expression::Concat {
-                    left: Box::new(new_l),
-                    right: Box::new(new_r),
-                    coerce_left,
-                    coerce_right,
-                },
-                ArkType::Bytes,
-            )
-        }
-        Expression::Neg64 { value } => {
-            let (nv, _) = rewrite_expression_concat(*value, scope);
-            (
-                Expression::Neg64 {
-                    value: Box::new(nv),
-                },
-                ArkType::Uint64Le,
-            )
-        }
-        Expression::Le64ToScriptNum { value } => {
-            let (nv, _) = rewrite_expression_concat(*value, scope);
-            (
-                Expression::Le64ToScriptNum {
-                    value: Box::new(nv),
-                },
-                ArkType::Int,
-            )
-        }
-        Expression::Le32ToLe64 { value } => {
-            let (nv, _) = rewrite_expression_concat(*value, scope);
-            (
-                Expression::Le32ToLe64 {
-                    value: Box::new(nv),
-                },
-                ArkType::Uint64Le,
-            )
-        }
-        Expression::ContractInstance {
-            contract_name,
-            args,
-        } => {
-            let new_args = args
-                .into_iter()
-                .map(|a| rewrite_expression_concat(a, scope).0)
-                .collect();
-            (
-                Expression::ContractInstance {
-                    contract_name,
-                    args: new_args,
-                },
-                ArkType::Bytes,
-            )
-        }
-        // Leaves and other compound expressions: no `+` to rewrite below the surface.
-        other => {
-            let t = crate::typechecker::infer_type(&other, scope);
-            (other, t)
         }
     }
 }
